refactor: Move mock LN adapter fixtures into fedimint-testing
